[Chore #161363562] refactored flask app creation logic to app/__init.py
